Move ExpressionEvaluatorSpec
