+++ core/src/test/resources/scala/EvaluateTest.scala
